fix(dashboard): handle opener errors and platform tags
