+++ dashboard/open_command_test.go
@@ -1,4 +1,4 @@
-//go:build !windows
+//go:build darwin || linux
 
 package main
