+++ dashboard/open_darwin.go
@@ -1,0 +1,7 @@
+//go:build darwin
+
+package main
+
+func openWithDefaultApp(target string) error {
+	return runOpenCommand("open", target)
+}
