--- dashboard/open_default.go
@@ -1,19 +1,0 @@
-//go:build !windows
-
-package main
-
-import (
-	"os/exec"
-	"runtime"
-)
-
-var runOpenCommand = func(name string, args ...string) error {
-	return exec.Command(name, args...).Run()
-}
-
-func openWithDefaultApp(target string) error {
-	if runtime.GOOS == "darwin" {
-		return runOpenCommand("open", target)
-	}
-	return runOpenCommand("xdg-open", target)
-}
