+++ dashboard/open_linux.go
@@ -1,0 +1,7 @@
+//go:build linux
+
+package main
+
+func openWithDefaultApp(target string) error {
+	return runOpenCommand("xdg-open", target)
+}
